Suppress conversion warnings when compiling with old gcc versions (#183)
* Suppress conversion warnings when compiling with old gcc versions


* Uncrustify


---------

+++ src/cpp/Cdr.cpp
@@ -2374,7 +2374,8 @@ void Cdr::xcdr1_serialize_short_member_header(
 
     make_alignment(alignment(4));
 
-    uint16_t flags_and_member_id = (member_id.must_understand ? 0x4000 : 0x0) | static_cast<uint16_t>(member_id.id);
+    uint16_t flags_and_member_id = static_cast<uint16_t>(member_id.must_understand ? 0x4000 : 0x0) |
+            static_cast<uint16_t>(member_id.id);
     serialize(flags_and_member_id);
     uint16_t size = 0;
     serialize(size);
@@ -2399,7 +2400,8 @@ void Cdr::xcdr1_serialize_long_member_header(
 {
     make_alignment(alignment(4));
 
-    uint16_t flags_and_extended_pid = (member_id.must_understand ? 0x4000 : 0x0) | static_cast<uint16_t>(PID_EXTENDED);
+    uint16_t flags_and_extended_pid = static_cast<uint16_t>(member_id.must_understand ? 0x4000 : 0x0) |
+            static_cast<uint16_t>(PID_EXTENDED);
     serialize(flags_and_extended_pid);
     uint16_t size = PID_EXTENDED_LENGTH;
     serialize(size);
@@ -2426,7 +2428,7 @@ void Cdr::xcdr1_change_to_short_member_header(
     assert(0x3F00 >= member_id.id);
     assert(std::numeric_limits<uint16_t>::max() >= member_serialized_size );
 
-    uint16_t flags_and_member_id = (member_id.must_understand ? 0x4000 : 0x0) |
+    uint16_t flags_and_member_id = static_cast<uint16_t>(member_id.must_understand ? 0x4000 : 0x0) |
             static_cast<uint16_t>(member_id.id);
     serialize(flags_and_member_id);
     uint16_t size = static_cast<uint16_t>(member_serialized_size);
@@ -2446,7 +2448,8 @@ void Cdr::xcdr1_change_to_long_member_header(
     {
         throw NotEnoughMemoryException(NotEnoughMemoryException::NOT_ENOUGH_MEMORY_MESSAGE_DEFAULT);
     }
-    uint16_t flags_and_extended_pid = (member_id.must_understand ? 0x4000 : 0x0) | static_cast<uint16_t>(PID_EXTENDED);
+    uint16_t flags_and_extended_pid = static_cast<uint16_t>(member_id.must_understand ? 0x4000 : 0x0) |
+            static_cast<uint16_t>(PID_EXTENDED);
     serialize(flags_and_extended_pid);
     uint16_t size = PID_EXTENDED_LENGTH;
     serialize(size);
